Refactor Promise to top level.
