+++ backend/src/api/NewsletterBrevo.ts
@@ -3,7 +3,8 @@ import * as SibApiV3Sdk from '@getbrevo/brevo'
 import { ContactForm } from '@prisma/client'
 
 import config from '#config/config'
-import { prisma } from '#src/prisma'
+
+import { SmtpEmailResponse } from './type/SmtpEmailResponse'
 
 const createBrevoInstance = (): SibApiV3Sdk.TransactionalEmailsApi => {
   const apiInstance: SibApiV3Sdk.TransactionalEmailsApi = new SibApiV3Sdk.TransactionalEmailsApi()
@@ -28,82 +29,64 @@ const createSmtpEmail = (
   return sendSmtpEmail
 }
 
-const sendSmtpEmail = (smtpEmail: SibApiV3Sdk.SendSmtpEmail, contactForm: ContactForm): void => {
+const sendSmtpEmail = (smtpEmail: SibApiV3Sdk.SendSmtpEmail): Promise<SmtpEmailResponse> => {
   const apiInstance = createBrevoInstance()
-
-  void apiInstance.sendTransacEmail(smtpEmail).then(
-    async (data) => {
-      // eslint-disable-next-line no-console
-      console.log('API called successfully. Returned data: ', JSON.stringify(data))
-      // code to store success goes here:
-      contactForm.brevoSuccess = new Date()
-      await prisma.contactForm.update({
-        where: {
-          id: contactForm.id,
-        },
-        data: {
-          ...contactForm,
-        },
-      })
-      return true
-    },
-    // eslint-disable-next-line promise/prefer-await-to-callbacks
-    function (error) {
-      // eslint-disable-next-line no-console
-      console.error(error)
-    },
-  )
+  return apiInstance.sendTransacEmail(smtpEmail)
 }
 
-export const sendContactFormEmail = (contactForm: ContactForm): void => {
-  if (config.BREVO_KEY) {
-    const smtpEmailToAdmin: SibApiV3Sdk.SendSmtpEmail = createSmtpEmail(
-      config.BREVO_TEMPLATE_CONTACT_BASE,
-      [
-        {
-          name: config.BREVO_CONTACT_REQUEST_TO_NAME,
-          email: config.BREVO_CONTACT_REQUEST_TO_EMAIL,
-        },
-      ],
-      {
-        name: contactForm.firstName + ' ' + contactForm.lastName,
-        email: contactForm.email,
-      },
-      {
-        name: contactForm.firstName + ' ' + contactForm.lastName,
-        email: contactForm.email,
-      },
-      {
-        email: contactForm.email,
-        firstname: contactForm.firstName,
-        lastname: contactForm.lastName,
-        content: contactForm.content,
-      },
-    )
-    void sendSmtpEmail(smtpEmailToAdmin, contactForm)
-
-    const smtpEmailToClient: SibApiV3Sdk.SendSmtpEmail = createSmtpEmail(
-      config.BREVO_TEMPLATE_CONTACT_USER,
-      [
-        {
-          name: contactForm.firstName + ' ' + contactForm.lastName,
-          email: contactForm.email,
-        },
-      ],
-      {
-        name: config.BREVO_CONTACT_REQUEST_TO_NAME,
-        email: config.BREVO_CONTACT_REQUEST_TO_EMAIL,
-      },
-      {
-        name: config.BREVO_CONTACT_REQUEST_TO_NAME,
-        email: config.BREVO_CONTACT_REQUEST_TO_EMAIL,
-      },
-      {
-        firstname: contactForm.firstName,
-        lastname: contactForm.lastName,
-        content: contactForm.content,
-      },
-    )
-    void sendSmtpEmail(smtpEmailToClient, contactForm)
+export const sendContactFormEmail = (
+  contactForm: ContactForm,
+): Promise<[SmtpEmailResponse, SmtpEmailResponse]> => {
+  if (!config.BREVO_KEY) {
+    throw new Error('No Brevo_Key defined could not send email')
   }
+  const smtpEmailToAdmin: SibApiV3Sdk.SendSmtpEmail = createSmtpEmail(
+    config.BREVO_TEMPLATE_CONTACT_BASE,
+    [
+      {
+        name: config.BREVO_CONTACT_REQUEST_TO_NAME,
+        email: config.BREVO_CONTACT_REQUEST_TO_EMAIL,
+      },
+    ],
+    {
+      name: contactForm.firstName + ' ' + contactForm.lastName,
+      email: contactForm.email,
+    },
+    {
+      name: contactForm.firstName + ' ' + contactForm.lastName,
+      email: contactForm.email,
+    },
+    {
+      email: contactForm.email,
+      firstname: contactForm.firstName,
+      lastname: contactForm.lastName,
+      content: contactForm.content,
+    },
+  )
+
+  const smtpEmailToClient: SibApiV3Sdk.SendSmtpEmail = createSmtpEmail(
+    config.BREVO_TEMPLATE_CONTACT_USER,
+    [
+      {
+        name: contactForm.firstName + ' ' + contactForm.lastName,
+        email: contactForm.email,
+      },
+    ],
+    {
+      name: config.BREVO_CONTACT_REQUEST_TO_NAME,
+      email: config.BREVO_CONTACT_REQUEST_TO_EMAIL,
+    },
+    {
+      name: config.BREVO_CONTACT_REQUEST_TO_NAME,
+      email: config.BREVO_CONTACT_REQUEST_TO_EMAIL,
+    },
+    {
+      firstname: contactForm.firstName,
+      lastname: contactForm.lastName,
+      content: contactForm.content,
+    },
+  )
+  const sendEmailClient: Promise<SmtpEmailResponse> = sendSmtpEmail(smtpEmailToClient)
+  const sendEmailAdmin: Promise<SmtpEmailResponse> = sendSmtpEmail(smtpEmailToAdmin)
+  return Promise.all([sendEmailClient, sendEmailAdmin])
 }
