+++ backend/src/api/type/SmtpEmailResponse.ts
@@ -1,0 +1,10 @@
+// eslint-disable-next-line import/no-namespace
+import * as http from 'http'
+
+// eslint-disable-next-line import/no-namespace
+import * as SibApiV3Sdk from '@getbrevo/brevo'
+
+export type SmtpEmailResponse = {
+  response: http.IncomingMessage
+  body: SibApiV3Sdk.CreateSmtpEmail
+}
