+++ backend/src/graphql/resolvers/ContactFormResolver.ts
@@ -2,6 +2,7 @@ import { ContactForm } from '@prisma/client'
 import { Resolver, Mutation, Query, Arg } from 'type-graphql'
 
 import { sendContactFormEmail } from '#api/NewsletterBrevo'
+import { SmtpEmailResponse } from '#api/type/SmtpEmailResponse'
 import { ContactFormInput } from '#inputs/ContactFormInput'
 import { prisma } from '#src/prisma'
 
@@ -13,7 +14,28 @@ export class ContactFormResolver {
   ): Promise<boolean> {
     const contactForm: ContactForm = await prisma.contactForm.create({ data: contactFormData })
     // code to send email goes here
-    void sendContactFormEmail(contactForm)
+    const contactFormEmailPromise: Promise<[SmtpEmailResponse, SmtpEmailResponse]> =
+      sendContactFormEmail(contactForm)
+    try {
+      await contactFormEmailPromise.then(async (data) => {
+        // eslint-disable-next-line no-console
+        console.log('API called successfully. Returned data: ', JSON.stringify(data))
+        // code to store success goes here:
+        contactForm.brevoSuccess = new Date()
+        await prisma.contactForm.update({
+          where: {
+            id: contactForm.id,
+          },
+          data: {
+            ...contactForm,
+          },
+        })
+        return true
+      })
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.error(error)
+    }
     return true
   }
